wip: pick conf from global telescope conf
